ENT-9941: Moved new connector factory to node-api (#7369)
